# docs/inji-certify-openapi.yaml
openapi: 3.0.1
info:
  version: v0
  title: OpenAPI definition
servers:
  - url: 'http://localhost:8090/v1/certify'
    description: Generated server url
paths:
...
  /.well-known/openid-credential-issuer:
    get:
      tags:
        - Well-known
      summary: Issuer metadata
      description: 'This endpoint provides metadata about the credential issuer and the supported credential configurations. The response structure adapts based on the version query parameter, allowing clients to discover issuer capabilities and supported credential types for different OID4VCI versions. The response will be as per the applicable OID4VCI specification, for e.g. for draft 13 of OID4VCI, the specification to be referred is https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0-ID1.html#name-credential-issuer-metadata-p'
      operationId: getCredentialIssuerMetadata
      parameters:
        - name: version
          in: query
          required: false
          schema:
            type: string
            default: latest
            enum:
              - latest
              - vd11
              - vd12
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialIssuerMetadataDTO'
...
  /issuance/.well-known/did.json:
    get:
      tags:
        - VC Issuance
      summary: Issuer DID document
      description: 'This endpoint returns the DID document associated with the issuer. This document contains the decentralized identifier and related public keys, enabling clients to verify the issuer’s identity and interact securely. The response will be as per the specification provided here : [https://www.w3.org/ns/did/v1](url)'
      operationId: getDIDDocument
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VCError'
      servers:
        - url: 'http://localhost:8090/v1/certify'
          description: Generated server url
      deprecated: true
...
components:
  schemas:
...
    CredentialDefinition:
      type: object
      properties:
        type:
          type: array
          items:
            type: string
          description: List of credential type strings.
        credentialSubject:
          type: object
          additionalProperties:
            type: object
          description: Map defining the schema and fields for the credential subject.
        '@context':
          type: array
          items:
            type: string
          description: List of JSON-LD context URIs.
      required:
        - type
...
    VCError:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
        c_nonce:
          type: string
        c_nonce_expires_in:
          type: integer
          format: int32
...
    CredentialIssuerMetadataDTO:
      type: object
      properties:
        display:
          type: array
          description: 'List of localized display information (such as name, logo, and description) for the credential issuer, used for UI presentation in wallets.'
          items:
            type: object
            additionalProperties:
              type: string
        credential_issuer:
          type: string
          description: The identifier (URL or DID) of the credential issuer. Used by wallets to recognize the issuer of verifiable credentials.
        authorization_servers:
          type: array
          description: List of OAuth 2.0/OpenID Connect authorization server URLs that can be used for credential issuance authorization.
          items:
            type: string
        credential_endpoint:
          type: string
          description: The API endpoint URL where clients can request credential issuance.
        credential_configurations_supported:
          type: object
          description: Map of Verifiable Credential level configuration information.
          additionalProperties:
            $ref: '#/components/schemas/CredentialConfigurationSupportedDTO'
    CredentialConfigurationSupportedDTO:
      type: object
      properties:
        format:
          type: string
          description: 'The credential format supported (e.g., ldp_vc, mso_mdoc, vc+sd-jwt).'
        scope:
          type: string
          description: The OAuth2 scope value associated with this credential configuration.
        claims:
          type: object
          description: Map of claim definitions for the credential (used for mso_mdoc and vc+sd-jwt).
          additionalProperties:
            type: object
        display:
          type: array
          description: 'List of localized display information for the credential (name, description, logo, etc.).'
          items:
            type: object
            additionalProperties:
              type: object
        order:
          type: array
          description: List specifying the display order of claims or fields in the credential.
          items:
            type: string
        doctype:
          type: string
          description: 'Document type identifier, used for mDL/mDoc credentials.'
        cryptographic_binding_methods_supported:
          type: array
          description: 'List of supported cryptographic binding methods (e.g., did, holder\_binding).'
          items:
            type: string
        credential_signing_alg_values_supported:
          type: array
          description: 'List of supported credential signing algorithms (e.g., ES256, EdDSA).'
          items:
            type: string
        proof_types_supported:
          type: object
          description: 'Map of supported proof types and their parameters (e.g., jwt, ldp).'
          additionalProperties:
            type: object
        credential_definition:
          $ref: '#/components/schemas/CredentialDefinition'
        vct:
          type: string
...